+++ source/VirtualScroll/VirtualScroll.test.js
@@ -499,4 +499,9 @@ describe('VirtualScroll', () => {
       expect(rowRendererCalled).toEqual(false)
     })
   })
+
+  it('should set the width of the single-column inner Grid to auto', () => {
+    const rendered = findDOMNode(render(getMarkup()))
+    expect(rendered.querySelector('.Grid__innerScrollContainer').style.width).toEqual('auto')
+  })
 })